Split Versions.kt into pieces
